Pass BrowserContextTimezoneTests
related to #499
